Moved StreetAddress object into the auxiliary package.
