+++ backend/src/main/java/gov/cdc/usds/simplereport/db/model/auxiliary/StreetAddress.java
@@ -1,4 +1,4 @@
-package gov.cdc.usds.simplereport.db.model;
+package gov.cdc.usds.simplereport.db.model.auxiliary;
 
 import org.hibernate.annotations.Type;
 import org.springframework.boot.context.properties.ConstructorBinding;
